[Search] Create search feature

+++ app/src/main/java/com/bku/appbooking/category/CategoryDataReceiver.java
@@ -74,7 +74,7 @@ public class CategoryDataReceiver {
                             String tilte = product_json.optString("tieude");
                             String price = product_json.optString("price");
                             String imgUrl = product_json.optString("image");
-                            long id = Long.valueOf(key);
+                            int id = Integer.valueOf(key);
                             products.add(new Product(id, tilte, price, imgUrl, "", ""));
                         } else if (object.get(key) instanceof String){
                             String value = object.getString("type");
